refactor: 抽取配置差异构建工具
